proper delta placement
